% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Funcion for title case the Earthquake's Location Data-Name}
\usage{
eq_location_clean(.data)
}
\arguments{
\item{.data}{is the dataframe that contains location names written in upper case}
}
\value{
a dataframe which contains the Eathquake data filtered required for mapping in a timeline the data and the Tittle Case Location
}
\description{
Funcion for title case the Earthquake's Location Data-Name
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes_data.txt.zip",package="capstone")
eq_location_clean(eq_clean_data(eq_data_read(filename)))
}

}
