% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Function that will use the GeomTimeLine Prototype Function required to Plot a Timeline with the Earthquakes of a given country}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  na.rm = TRUE,
  position = "identity",
  stat = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings created by aes}

\item{data}{is the dataframe that contains the Earthquake's data}

\item{na.rm}{will hepls to remove the NA values from the data frame}

\item{position}{position adjustment functio}

\item{stat}{The Layer's statistical transformation}

\item{show.legend}{layer's legend}

\item{inherit.aes}{will indicate the default aesthetics overridng}

\item{...}{layer's other arguments}
}
\value{
In a plot an Earthquakes timeline which contains all Earthquakes of a Given Country or List of Countries between a set of dates
}
\description{
Function that will use the GeomTimeLine Prototype Function required to Plot a Timeline with the Earthquakes of a given country
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes_data.txt.zip",package="capstone")
eq_location_clean(eq_clean_data(eq_data_read(filename))) \%>\%
dplyr::filter(datetime >= "1980-01-01"
& datetime <="2014-01-01"
& COUNTRY == c("MEXICO","USA", "JORDAN"))\%>\%
ggplot() +
geom_timeline(aes(x = datetime, size = EQ_MAG_ML, colour = DEATHS, fill = DEATHS))
}

}
